Initial Commit!
So happy to be trying this out now! This is gonna be a cross-platform NBT editor that runs on any device or browser, and it's built with vanilla JavaScript, ES Modules, and Web Components. Some of these may change with time, like possibly moving to TypeScript, using a UI framework, things like that. But, getting started with the actual project, this is how I'm going to start things out.

Got JS+JSDoc IntelliSense fully working, got my Node Modules synced up with my tsconfig, and it's already set up to work offline. Now I can get things working for the actual app!

Couldn't think of a name for the editor off the top of my head. The ones I though of are either already taken, or not very creative. I thought of the name Dovetail a few years ago, when I was working on my Linear Logic resource pack project. That one was built to allow you to run Java Edition resource packs directly in Bedrock Edition, with no conversions. I didn't end up continuing that project, so it doesn't have a GitHub repo or anything like that. It used to be on MCPEDL when it was out too. I still have the project files for it, so maybe I should actually make a repo for it, at least for archives' sake.

Anyways, Dovetail was one of the contending names for that project. I ended up going with Linear Logic as I just mentioned, so I was still fond of Dovetail for a project name. I looked into the NBT spec today to find anything that may stand out there as a neat name. One I thought of was 'Notchian', which I think I saw somewhere a few months ago while working on NBTify. It was in the docs, and it was in reference to the "original Notchian specification" for the NBT format. I did like that one, but then I remembered my Dovetail naming idea, and I realized that that one was similar to the name 'Notch' itself! Kinda neat. And when I Googled the definition for dovetail once again to get a refresher, it mentions 'notches' right there, so that was cool.

And thinking of it now, I really like the concept of how a dovetail-joint graphic could look for an app icon, now I can't wait to design one! That kind of looks like a fancy way to represent NBT to me, too.

Oh yeah, and I thought Dovetail was a neat name for Linear Logic back then too, because it kind of ties the resource pack formats together just like a dovetail-joint looks. Linear Logic kind of took a similar icon path, to where it was a corner puzzle piece.