Import Map Module Resolution
Tried out setting up module bundling (with Vite), but I think I still like this bundle-less setup. I wish this could just resolve back to CDN URLs when compiling down from TypeScript, but doesn't seem like that's a thing people have tried to do yet. Maybe I should make something to do that!

This current setup is polyfilling the new Import Map module resolution spec, which is currently only supported in Chromium and very modern Firefox. This shim fills in a lot of the missing gaps to make it work about 95% of the time across all browsers. It shims the module loader to respond with a Blob import of the resolved module. It's super clever!

https://github.com/guybedford/es-module-shims