Workflow Repo Paths
Realized I didn't add the correct handling for where the repo is cloned from, I was supposed to add my own info in there, haha.

Not gonna lie, I've been using ChatGPT to help me out with learning these new tools, as they seem to be more confusing for me to understand than coding sometimes, geeez man. I see how they can speed things up, but they definitely take a while to fully dive into using in your own coding setup. Once you get on board though, it really is another level of cool. You just have to take that first leap! I guess I'm trying to do that with ChatGPT too, and it has shown itself to be very helpful to me already so far. I think my favorite thing about it is it's understanding in context for things you are asking it. It's less about keywords, and more about providing it with real context that it can break down and try to understand what you're trying to do. That's a really big thing to be able to break down. My thought about it is that it's still the same data that we're working with already (everything on the Internet), but it's a much more refined interface to be able to look at it through. The data's already out there, and now we have another tool to be able to look though, compare, and discern what we already know and have learned (as people, I guess I'd say). Of course the data on the Internet doesn't perfectly depict all of the information out in the world, but that's not ChatGPT's issue, that's because of what we've all decided to put up on the Internet in the first place. It also could be ChatGPT's fault for both parsing that Big-Data inaccurately, so I guess it's a bit 'lossless' on both sides, at least to how I'm thinking of it. If we as people can discern what we get out of our conversations with projects like ChatGPT, we can then *ourselves* also get the big picture from what it responds, since that's exactly what it's (GPT) doing also. I think this makes it "reliable-ish" in a way. I think navigating Wikipedia this way is also very smart. For how much data there is on a given resource you're looking at, I'd like to think that some majority of it is most likely credible. If you understand and accept that you don't know for sure though, than you can got to multiple different resources to try and see what each one represents. Then you yourself can decide how you want to take those conclusions.

Not sure what that rant was, or where that came from. That was awesome! XD