Format Options Component
Moved the setup for the Format Options dialog and form into it's own Web Component. Not sure how I like it set up like this, but it's definitely going in the right direction.

Finished this a few hours ago, but I got sidetracked watching an awesome interview with Mike Keneally about his solo catalog. Then I took a break for lunch :)
https://www.youtube.com/watch?v=CKNIe347j5o