FormData Object Generator
Dang, this is still ugly. I'm curious if this is where people start to use React for things. This is a bit hairy just to have a mirror copy between my form data and a simple JS object. Too bad I don't know how to set up React real quick, that may have to be one of the next things I look into :O