Vite Bundling
Got Vite working! I've been using `live-server` for everything up until now, so this will be a significant change to how I've been doing things. It's gonna get hard to learn how to make things differently using this more-commonplace setup that a lot of people have been using.

I want to explore diving deeper into making a more modern development setup, as my stack up until now has been fairly vanilla, and manually managed. I want to learn more about the different tooling and frameworks out there, as I have kind of shied away from it, either because of my young experience in working things like it, and because I haven't had the most capable or stable dev enviromnents to be able to heavily depend on. Now I can fully rely on my new machine, and I can also rely on using Git and GitHub for things, which was a bit out of my reach for some time. Now I think I understand it a bit better, and I understand the scope to what each thing does. This has been the same for TypeScript too.

As of this commit being pushed, this will be on a development branch of the project, as I have to do a few more things to get it all ready for production use. I want to set up a proper GitHub Action, which more appropriately manages the build of the app. It's a bit messy how I have it set up right now, it runs the GitHub Action to build the app, pushes those changes to the `github-pages` branch, which then causes the class GitHub Pages action to build the 'Jekll' site from there, and publish that to the world. So there's an extra step in there I want to remove. I have been hearing that you don't need a dedicated source branch for the deployment of the app anymore, and you can simply just push the build to host itself separately. I really like that! Having a branch just for the build of the site has been my main holdup of wanting to have a build step for my sites, I didn't want to track build files as much as possible anywhere in the repo. It seemed a bit bloated to me. Not that it's much of a performance hindrance for a small project, but that's the case only for while the project is small, and I didn't want things to get out of hand befor it got messy and turned into something I had to fix. Now that's not an issue :)

Listening to Wing Beat Fantastic!

Started playing 'Elektric Band', which made me think of Herbie Hancock Rockit, remembered how good that song is. That then reminded me of XTC 'Making Plans for Nigel', which then led me to find that the Andy Partridge in that XTC band was the same Andy Partridge in the 'other' XTC band I thought I heard yesterday in the Mike Keneally interview. Really awesome crossover, now that I know both sides of those! I knew Mike Keneally and 'Making Plans for Nigel', but didn't realize that Andy Partridge was in 'Nigel'. Made me want to listen to 'Wing Beat Fantastic' again! I was curious to hear if there's any musical crossovers that I can hear out of there. Not sure how much of the crossover I can hear yet, but it helped me try to listen for things a bit better now. I don't know Partridge's writing other than 'Nigel', so I don't know how his influence sounds yet. I really like the writing now that I know who he is! Not sure how that works, but I can hear it a bit different now. :D

I love the intricacies of music!

I listened to Storm Corrosion again before this, and I realized that I heard my guitar teacher's influence of Opeth in his own writing, because I heard one of his albums in what was likely Mikael Akerfeldt's part of the opening song on Storm Corrosion. So, if that didn't make sense, I'll put it into other words. Opeth sounds like Opeth because of Mikael Akerfeldt. My guitar teacher likes Opeth, so one of the songs on his album sounds like Opeth. By the transitive property, Mikael Opeth's writing (I KEEP ACCIDENTALLY SAYING THAT XD) on Storm Corrosion reminds me of the writing on my guitar teacher's album. Very awesome stuff! It also happened a little while ago where I can hear Mikael's influence on Steven Wilson's writing with Porcupine Tree, namely parts from Fear of a Blank Planet and Deadwing sound like parts of Blackwater Park. That's even more of a twister though, as SW took part in working on Opeth's Blackwater Park too, so it gets really confusing on who influenced who's sound there. aaaah! haha

Had to take a break while writing the first sentence of the previous paragraph, went on a walkies out in the woods with mom and the dog. We crossed a river stream, and it partially inspired me for a new song, which didn't start working out as I had wanted it to. The recordings of the nature scene sound great, and gives a great atmosphere, but I can't quite figure out how the song should go. I was going to try and encorporate a Herby Hancock/XTC 80's drum vibe on top of it, to make it confusing, but nothing could come out of it. I think I'm not down the right avenue yet. Gonna try again after writing this message, and the rest of the coding I'm gonna do. Gotta finish listening to Wing Beat Fantasic, too!

Note to self: This isn't available on my Git Blog yet (it's not set up for anything still, yet), but these are the great kinds of commits that I'm looking for to be able to share with others through a blog-like form. Having experience with a build step would be great for getting that set up!

asta la vista, mis amigos.
pas.

-finished this message at 6:24, haha