Bundle Deploying
Still a bit confused about how to manage all of this, and I still can't get some parts of it to work either :(

For some reason I can't get `vite preview` to run the app consistently on my own machine, so I'm not sure I want to push the site using the bundled version, if I can't try it out as-is on my own machine.

I am definitely very close to getting it working I think, though! Now I have an example script from the Vite team on how to deploy the build to GitHub Pages too, which I can use for my existing production setup with TypeScript! That's super cool.

https://vitejs.dev/guide/static-deploy.html

Gonna keep this working update in it's own branch for the meantime. Gonna further look into getting all of this working, as I really want to have a simple to use *and* understand dev setup. For myself personally at least, it feels a bit overcomplicated in the fact that so many parts of the dev setup are all tied together, and it feels like that's making it harder to get setup just right, and to figure out how it all works together.

It's still a bit surprising how hard it is just to get a built site up with vanilla code (TS, of course though), and not preprocess every single part of the codebase. Maybe that's what you're supposed to think like, it seems a bit foreign to me still. To have just learned things like Web Components and ESM, it feels a bit weird to use `import` statements for small things like image URLs. Why not just add it to your HTML then? I wish I could use the same degrees of separation as before, just with better language features and support.

And geez, why can't TypeScript and Service Workers just WORK, gosh dangit!!! aaaargh.

All steps back are learning experiences though, so this negativity is a good thing! I'm learning more :) (Finished Elektric Band 'All Love' right as I was typing that, it made me feel really good inside :) )