Strict Mode Dialog
Bringing the latest NBTify changes over to Dovetail! The first of the features is the new handling for strict mode, which allows for files that have trailing bytes at the end to be opened, if specified by the user. Since I want to ensure no data to be lost as much as possible, I made it visible to the user when the file they are trying to open has trailing bytes at the end of the file. Since some files have this intentionally it may seem (LCE?), I'm making this optional to decide which way you'd like to handle it. It's possible that there actually is an error, and NBTify isn't reading all of the data correctly, or that the file just actually is a little bigger than the NBT data itself that it holds. I think that may be what's going on for the LCE player files, and other NBT libraries appear to skip over the rest of the file, even though it goes as unread. So this seemed a bit unsafe to me, as with NBTify's format auto-detection, you can't always be sure that the extra bytes not being read is intentional or not, because it might just be reading it in the wrong format. I have seen it where NBTify can successfully open a file early using a different format, when there are still a lot of bytes left, and they actually do hold what the NBT data should be. So this could be dangerous to the user's data if this isn't managed safely and made visible during the reading steps.

Noticed that `.dat_old` files couldn't be opened, so I added them as a file association too.

I tried adding NBTify's new Deflate-Raw handling into the Format Options menu to allow Deflate-Raw inputs and outputs, but I tried opening a file I wrote with Deflate-Raw in the editor, and realized that NBTify currently doesn't have a way to auto-detect Deflate-Raw, which means Dovetail also doesn't have a way to open files compressed with that. It will simply error out and not be able to open the file, because Deflate-Raw is something that it knows how to handle, but isn't something that it can discern against from the other formats yet. So in JavaScript land, you can specify that it's Deflate-Raw, and it will open perfect, but Dovetail can't because that relies on figuring out the format for you, so you don't have to tell it how to open it.

Skull Bubbles is a great song.