Proper Vite Build!
I think I actually got Vite to work this time! It was still kind of a mess unfortunately, I think it's just because I don't quite know where the differences are in terms of how bundlers and static sites work. I'm still very static-minded in terms of what I think the build output will look like (I kind of pictured that everything would be symmetrical in terms of the folder structure for the project source, but it seems to be more isomorphic than I thought).

I think the biggest thing I learned here today is that the `/public/` folder has more significance than I thought. It's where you place your files that you want to go into the project unchanged, as-is. Finding this out was extremely helpful, and I don't know how this isn't pointed out explicitly anywhere. I know it's probably just common knowledge by now for devs out there, but I'm kind of baffled that I've only found a single source that describes this is how it works. It's like a chicken or the egg thing, the original idea for this must have been implemented a long time ago? Knowing this now makes much more sense, I thought it was just a commomplace thing to put files in there, I didn't know it was a requirement in terms of the resolution process for the app's build step, the more you know!

And the other one that seems very simple of me to miss, is that you have to build the app with `vite build` (creates the `./dist/` output) before you can run it with `vite preview`. I thought that preview would just do the same thing as `vite`, which does everything 'magically' in a way, but I do like that it's more explicit in that way for testing the actual live build output. I think it's good that you are able to see what Vite actually produces from your source, which seems much safer to debug since you have full access to seeing what the build output looks like, rather than only what is visible to you in the browser.

https://www.reddit.com/r/webdev/comments/z6dsoh/can_someone_explain_to_me_what_the_public_folder/
https://github.com/vitejs/vite/discussions/4572
https://vitejs.dev/guide/cli.html

https://dev.to/wtho/custom-service-worker-logic-in-typescript-on-vite-4f27 (Didn't really help in terms of the config part, but it helped my brain click in terms of module resolution in finding the worker script with `navigator.serviceWorker.register()`, which I think ended up kind of working when you resolved things to `./src/service-worker.ts`. Don't quote me on that though, I think that might not have worked for `vite`, but it does for `vite preview`, which was mentioned in a comment further down after the article)

I do have to say, I like this more than the `live-server` dual-terminal setup I had before. That was a little easier to setup, but this is a bit more streamlined in terms of a dev process. For my previous setup, I'd run `tsc -w` in one terminal, then `live-server` in another, and then open that in the browser. Essentially, a lightweight TypeScript dev setup! That's what I'm using for STE at the moment too. I imagine once I get the hang of this setup though, I will move my projects over to bundling instead, since this is a bit more modern, and with GitHub Pages now supporting artifacts, it doesn't matter that the site source has to be the actual source code! Yay. And I think this is the more 'professional' way to build a site, if you wanted to say that. Gotta say, I'm liking how this is going :)