Svelte Testing
Started looking into Svelte yesterday! It seems really cool, I like it's vibes (I don't know what I'm doing here yet though, haha). Got interested after watching the 'Rich Harris Teaches The Primeagen Svelte' video. It's neat that pretty much all of it's features are built around the compile step, so you don't have to ship nearly as much JS to get it set up in your project (only the parts that your code relies on). Coming back to commit these changes though, I'm not quite sold on the dev setup for it, as it relies on a lot of tooling you have to rely on more separate tooling to set your project up for it. I like TypeScript in that you use it for everything, and what it builds out is what you run, and it doesn't look too much different from the source, because of the 'all JS is valid TS' correlation. I guess for some aspects of Svelte this is also the case (going between HTML, CSS, and JS (TS)), but I'm not sure how I feel now about actually coupling all of those together into a single file, which is solely for the build step to handle. I guess TypeScript is the exact same way, but I'd say less so because of the mostly-symmetrical nature it has. For Svelte components, the syntax 'looks' a lot like the other languages, but it does a lot more behind-the-scenes work to your initial source, so what you get out likely isn't as close to the original source (This is good for the performance enhancements I mentioned earlier, but this way of building things is still a bit foreign to me, at least to know how to structure things myself).

Got help from making a demo project off to the side of this one, which helped with figuring out what ways to set up Svelte are. `npm create vite@latest`