Deps + DOM Referencing
While sticking with a semi-vanilla stack for the time being, I think I don't actually mind using `declare var` for ID'ed element references. I like it a bit more for keeping track of what the HTML looks like, rather than using constants for it. I might bring STE back to this approach too. It's more of a type-level thing anyways, since they are defined on the global object itself implicitly.

I added `autocomplete` to the `<input>` because DevTools was suggesting it should be present to help with accessibility.

That also reminds me, I also still have to look into accessibility for Num Text and Menu Drop.