Logo Render (Blockbench)
Wanted to look into making a nicer render for the app's logo! Blockbench was the first to come to mind. It isn't screenshotting as nice as I'd like it too though, so I'm going to try a second round at this to get it right.

To actually render the model accordingly, I used the 'Isometric Right (2:1)' view option provided by the Minecraft Title Generator extension.

Couldn't remember if the repo at some point included the original dovetail joint image I used to create the current logo, so I relocated it online, and added it here again just in case. I just did some small paint bucket fills on the original image to get the current logo.

Wait a minute, why am't I using Figma for this?? Forgot about that lol, haven't used it in a while I guess! That's a great program too. That will be where I try next, as Blockbench isn't the best suited for logo generation, it isn't really meant for vectors and such. I mostly just wanted it for the 3D-rendering part.