SNBT Char Escapes + Deps Update
Bringing the updates from NBTify which fix handling of escaped characters while editing with SNBT. I was going to bring these changes to the main branch since that's still currently using SNBT for editing, but I decided to bring them here since I want to move this branch along first, rather than merge it later. I want to have both GUI and SNBT editing available, so I'm going to focus editing on this branch until it's in stable, rather than leaving it to the side and bringing some changes to main, and some to here. This is the future of Dovetail, so I'm gonna push this as the main focus to get it ready! It's a heck a lot of the way there already! Just have to add `<dialog>` value editing support next for it to be fully functional.