Svelte: The Deep Dive
Gonna do it. If I'm gonna start using Svelte, I have to start using it first. That's the best way to put it!

Took some of the setup from `npm create vite@latest`, and brought it to here with some changes to fit it into Dovetail's styles and formatting. Not sure about all of the things it changes yet, but I'm gonna have myself not worry about that for the time being, as I think that's what holding me back about it.

Here's to learning new things!