Semver npm
Was going to bring the app over to Svelte once and for all, but I'm not quite there on understanding how to use/implement things with it quite yet. Gonna keep going with the current vanilla model and Web Components for the time being, in favor of faster development on Dovetail while I'm not quite ready for Svelte.

Learned about the different between `npm install x y z` vs `npm upgrade (update) (--save)` to upgrade npm dependencies. I thought the first was what I wanted, but now I understand you want to allow dependencies of your package to fall between a range of versions that your package supports. So bringing the dependencies up to their highest releases in the `package.json` is less helpful to maintainers below, because it shrinks the range of available versions that they could depend on, if neighboring packages to yours require a different version of that same dependency. So having a wider range of supported dependencies is something you *do* want!

Going to start following Semver and proper dependency management from hereon out, now that it makes a bit more sense. So rather than running `npm install x y z` to update everything, I will just use `npm upgrade`, without the `--save` modifier, which will allow the dependency range to help with semver calculations across packages :)

https://jh3y.medium.com/how-to-update-all-npm-packages-in-your-project-at-once-17a8981860ea
https://stackoverflow.com/questions/49270776/npm-semver-versioning-updating-a-package-with-a-caret
https://docs.npmjs.com/cli/v8/commands/npm-update