Checking out Svelte Again
I want to better understand how to use Svelte, so I'm coming back to this project again to try things out, because I have had the most progress so far in this one.

I want to redo STE too now that I think about it. It might be nice to restart STE with Svelte as a base!