TS Service Worker!
Brought this over from my recent update to STE, which adds the same thing.

I'm not sure if I should/can keep the top-level `await` use of `navigator.serviceWorker.register()`, should I just let that do it's own thing in the background? I tried looking online to see if awaiting SW registration on the front end with top-level await is a bad idea, but the results more just came up with whether Service Workers themselves should allow for top-level await or not. STE currently lets it fall off of the await stack, I had to change it to that because it didn't work for some reason, I can't recall exactly. But it did freeze up the whole site if I remember right, because it was awaiting for the Service Worker to register, but it never revolved *or* failed, so it just got stuck. It wasn't frozen because the performance was too much to handle, but rather because it *had* to resolve for the next script calls on the page to continue running.