Modern Dovetail!
Haven't moved these changes to the main branch or production in a while, I think they're about safe now, so I think I'm ready to move things over.

Not quite a 2.0 release technically, but the numbers are there, and I want to work on making Dovetail have a better app experience, and I think moving onto the next stage is part of what that entails. Hence, things like #6. I think I can also start looking into more features like fancier styling, UX, animations, and things like that. And eventually I want to get SNBT syntax highlighting (and hopefully linting/syntax error validation) down the road as well.

In a few spur branches of the project, I tried integrating with Svelte and React, but I don't think my own skills are ready for that kind of step yet. Nor do I think Dovetail is either. Embedding all of React into the app made the bundle size a bit bigger, and it didn't add anything new that the app can't already do. I want to add a framework in when it drastically helps lesten the load of the app's own weight under it's own implementation. I guess that's where we're at with STE, and I don't really want to touch that craziness as of yet, so maybe I don't want to let Dovetail get to that point either. But I think that could also be inevitable, no matter how you choose to build the app. I guess it depends on how you build the app, not what you build it *with*. If you manage things appropriately, I don't think it necessarily has to get out of hand. You can make a total mess out of any tool someone hands to you, be it a framework or a library. You just have to know the most applicable ways to make use of it! Wow this sounded very fancy lol.