Keyboard Shortcuts
Want to get keyboard shortcuts working, I just need to figure out a nice way to manage shortcut handling. My only active implementation for shortcuts is in STE, and it's very spaghettish. This is my first thought on how I can handle things, using an enum and a switch statement to check each state a bit better, and to also more explicitly these at the type level.

I want to use a similar pattern to that of what Electron uses for it's accelerator pattern, I just haven't known how to approach managing that quite yet. Trying an enum and Set for this actually seems to be a good idea maybe! Now I just need to set up my own platform-unique handling for things like differing between Win-Linux, and Mac-like.

https://www.electronjs.org/docs/latest/api/accelerator

#6