NBTTree Accessibility
My current setup for my NBT tree renderer was using `<label>` elements, but they weren't associated with any `<form>` elements, which Chrome DevTools wasn't a fan of. I looked into what I should do to fix it, and I currently don't have something that it can be associated with. I then looked into what I should replace it with, and I think the next option I wanted to go with was the `<p>` tag, but that involved having some extra styling that I didn't want to account for with my current styles, and I'm still not sure if it's quite the right choice of element for that section of the UI. Going to go with a `<span>` for the time being, as it hast the most similar style to what I had before, and I think it possible could be more accessible to what kind of element I'm intending it to be anyways.

I'm thinking about trying to move things over to React or Svelte again, but we'll see.