SolidJS
Going to do it I think!

I was hesitant about both originally the React Vite plugin (I didn't fully get what it was for, but now it makes more sense), and I am currently for the SolidJS one too. It's what allows for hot module reloading, which is more specific than I thought previously. I'm used to when you save a file in your app, the whole page reloads. When you are building with a reactive framework though, the state of the DOM is managed by itself, so you can actually keep track of where you are currently, and still re-render the static things outside of the reactive parts. Very cool! I was going to leave this part out since I didn't want the extra things to install and download, but if I'm going to convince myself to move over to the framework train, I think the only way to do that is going to be with the most streamlined setup that people tend to recommend.

Looks like the plugin uses Babel and it's children dependencies, hence why there are so many things to install with it. If another version of it is made with SWC though, then you can instead use that standalone, which I think I remember hearing is full native Rust, which would probably be not only a smaller, but self-encapsulated dependency.

https://www.reddit.com/r/solidjs/comments/18wcv7n/is_there_an_swc_vite_plugin_for_solid/
https://stackoverflow.com/questions/16661982/check-folder-size-in-bash
https://www.reddit.com/r/node/comments/4z48e2/is_it_normal_to_have_a_100k_files780_mb_in_the/ (classic :p)
https://www.reddit.com/r/node/comments/121qmjb/do_you_actually_care_about_the_size_of_node/ (current day)

This is my very much my take as well, from the link above:

> It’s less about the size of node_modules to me, and more about installing way too many dependencies that can easily be done with node builtins, and now require frequent updating and dependency management.

It's why I like to try building my projects using no dependencies I think. It's all your own scope of things to worry about. If something isn't up to your own standards, it's both your own fault (lol), and you have the access to fix it, unlike needing to ask the parent repo to adjust it, or to submit a pull request, which doesn't always get upstreamed (I want to do more of this, I can do better at helping other repos by upstreaming rather than forking).