Don't Give Up! Give a Solid Effort, lol
Making progress! Was losing the flame a bit when things started to not work like I thought. Then I just looked a little closer, and realized where I goofed in a few spots. This is great! I love it! Thank you SolidJS :D

I think the best part so far is the global state, not needing to pass things down to child components.

Wow, with new ears from having been listening to new music, I forgot/didn't realize *just* how great "Contraire de la Chanson" really is. This album has so much depth in so many ways. I here past and future Marco ideas in multiple spots! It feels like a very close to the heart album of his. If I ever get to talk with/meet him in person, I definitely want to talk about this album a bunch.

I'm so glad my music ears have been broadening the last few years, it makes coming back to music you loved the first time, actually even better I think. Not sure what that is, but fresh ears are the best for an album you really like.

https://superuser.com/questions/1750645/lossless-extraction-of-audio-from-webm-and-mp4-files (Further things that helped with getting archived songs that aren't available anymore, how else can you get them if they aren't in print? 😓)
https://stackoverflow.com/questions/65511370/webm-vs-opus-quality-youtube-dl-context
https://www.reddit.com/r/youtubedl/comments/11qkpwj/how_to_download_best_audio_quality_for_selfusage/
https://github.com/yt-dlp/yt-dlp?tab=readme-ov-file#post-processing-options

https://docs.solidjs.com/guides/how-to-guides/comparison/react
https://dev.to/this-is-learning/components-are-pure-overhead-hpm
https://www.solidjs.com/docs/latest/api#createroot (I thought using `render()` or `createRoot()` was causing my issues in the previous commit, but it was just my implementation that was a bit wonk. Yay! I was worried it was something about the library I didn't get. It was just my derp, instead)
https://www.solidjs.com/docs/latest/api#render

This was the misunderstanding that I accidentally missed here. You want to make sure to pass the accessors themselves as props, rather than raw values. I really like this distinction, and the fact that the creator of Solid likes this raw values, vs reactive concept, is very hopeful for me, that is a big distinction I like to make as well. Having understood the need to pass accessors, now things make much more sense, and I really like how it's explicit at the type-level as well, since it's either reactive, or it's not. Unlike in React or Svelte when you use it like a constant value, but it works like a reactive one. That less obvious behavior is a bit confusing I feel, at least when you haven't learned to work enough in that model for it to become natural. I'm hoping working with this kind of setup in Solid may be a nice way to level up my reactive prop knowledge, and to get React under my belt, having made the distinction in this library.

https://www.reddit.com/r/reactjs/comments/tsx8hw/solidjs_devex_compared_to_react/ (This is the one with the discussion from Ryan_Solid, great!!)
https://github.com/solidjs/solid/discussions/749
https://www.solidjs.com/tutorial/flow_switch
https://www.solidjs.com/tutorial/flow_show