Render Learning
Trying to figure out where you need to use reactive accessors, and where you don't have to. I like to make variables that allow me to use more explicit control flow, which TypeScript is a bit better with, so while I do like the look of the `<Show>` component, it doesn't completely work with TS's inferences. So I think I'm going to stick with the `&&` syntax for now at least, because it allows for me to do more fancy validation control flow checks, which aren't completely possible with `<Show>`.

Forgot to mention in the previous commit that I found a bug that the auto unfolding for non List or Compound tags wasn't explicit enough to match the other list-based tags. Now the only tag that will auto-expand when opening the Tree View is the root tag, be it a List/Compound.

While having just tested that root List tags work still, I just discovered that the Tree View doesn't completely fully re-render without toggling the view. Gotta figure out what the issue with that is.