SNBT Tree View String Escaping
I plan to eventually better implement this from NBTify itself. I don't want the root `stringify()` function to accept any tag kind (for validation concerns), but I think it would be nice to export the `SNBTWriter` methods somehow as individual functions instead. Maybe the same can go for `NBTWriter` as well, if that is needed for some reason. I haven't exposed them as part of the public API as of yet because I haven't thought of a reason for needing to do this, relative to regular NBT/SNBT parsing or handling, but it does make sense for something like an application like Dovetail, where it deals with parts of the NBT tree individually for the UI rendering. It could also be useful for the Tree View UI if I add support for editing things from there, as well. The dialog box would ask you for what you want the new value to be, and maybe it could just be an SNBT editor for that value on it's own. Might be simple to understand, and work with the types, rather than needing to provide specific GUI buttons for it. Maybe both is the better option though, seems like the GUI route would be more user-friendly out of the gates.

Svelte 5 makes me interested in it again! It's more similar to that of what Solid has going for it, as well. Not sure about single-file components though, I think that might have changed with that version too as well though. The CSS handling and scoping for that is indeed really nice, but not at the sake of allowing you to do less things implicitly out of the box. I will probably come around to it eventually though, it is really nice! I don't like to get hindered by nice abstractions though, that's my main holdup with things I think, when learning new tech and new stacks.

https://www.youtube.com/watch?v=R615rlP8Bvs